requirement.txt
gruut
anyascii
inflect
MeCab
num2words
pandas
librosa
transformers
pytorch-lightning
sklearn